change the auth methode from jwt to password and email in the deletemyaccount

+++ controllers/storage.js
@@ -95,19 +95,29 @@ class Storage {
     return uploadBlobResponse;
   }
   async removeUploadedFile(url, containerName) {
-    console.log(
-      "🚀 ~ file: storage.js:98 ~ Storage ~ removeUploadedFile ~ url",
-      url
-    );
-    //it remains 7 days by the way but the link will be unuseable
-    this.containerClient =
-      this.blobServiceClient.getContainerClient(containerName);
+    try {
+      // console.log(
+      //   "🚀 ~ file: storage.js:98 ~ Storage ~ removeUploadedFile ~ url",
+      //   url
+      // );
+      //it remains 7 days by the way but the link will be unuseable
+      this.containerClient =
+        this.blobServiceClient.getContainerClient(containerName);
 
-    const blockBlobClient = this.containerClient.getBlockBlobClient(
-      blobNameFromUrl(url)
-    );
-    // return new Promise((resolve, reject) => resolve());
-    return await blockBlobClient.delete();
+      const blockBlobClient = this.containerClient.getBlockBlobClient(
+        blobNameFromUrl(url)
+      );
+      if (!blockBlobClient.exists)
+        return new Promise((resolve, reject) => resolve()); // if the img doesn't exists
+      // console.log(
+      //   "🚀 ~ file: storage.js:117 ~ Storage ~ removeUploadedFile ~ blockBlobClient.exists",
+      //   blockBlobClient.exists
+      // );
+
+      return await blockBlobClient.delete();
+    } catch (e) {
+      // console.log(e);
+    }
   }
   async validBlobLink(imagesUrl, containerName) {
     let isValid = true;
